+++ mail/proton-mail-uniffi/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased] - 2024-00-00
 
+### Fixed
+  - Restored paginator `has_next_page()`
+  - Login after log out
+
 ## [0.11.51] - 2024-09-25
 
 ### Changed